documind-backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0

python-multipart==0.0.22
python-dotenv==1.2.1

openai==2.21.0
httpx==0.28.1

PyMuPDF==1.26.7